+++ src/plugin/gui_framework/style.hpp
@@ -86,6 +86,7 @@ enum class Colour : u32 {
     Green,
     Red,
     Highlight,
+    Blue,
 
     Background0,
     Background1,
@@ -133,6 +134,7 @@ constexpr auto k_colours = [] {
             case Colour::None: result[i] = 0; break;
             case Colour::Green: result[i] = ToAbgr(0x40A02B); break;
             case Colour::Red: result[i] = ToAbgr(0xD20F39); break;
+            case Colour::Blue: result[i] = ToAbgr(0x1E88E5); break;
             case Colour::Highlight: result[i] = Hsla(k_highlight_hue, 93, 78, 100); break;
 
             case Colour::Background0: