+++ changelog.md
@@ -12,6 +12,7 @@ For instance: 0.0.1. Don't change the headings.
 ## 0.0.8-alpha
 - Brand new instrument browser supporting tags, search, and filtering by library
 - Brand new impulse response browser featuring all the options of the instrument browser
+- New save preset dialog with author, description, and tags
 - Fix instrument left/right and randomise buttons
 
 ## 0.0.7-alpha